用户模块

+++ flag-api/src/main/java/com/luoys/upgrade/flag/api/bo/UserBO.java
@@ -24,11 +24,11 @@ public class UserBO {
     /**
      * 1--管理员；2--普通账号
      */
-    private Byte type;
+    private Integer type;
 
     /**
      * 1-正常；
      */
-    private Byte status;
+    private Integer status;
 
 }
